chore: move transaction serialization functions into serialization dir
