+++ libs/coin-framework/src/index.ts
@@ -1,0 +1,5 @@
+import * as serialization from "./serialization";
+
+export default {
+  serialization,
+};
