+++ libs/coin-framework/src/serialization/index.ts
@@ -1,3 +1,4 @@
 export { fromAccountRaw, toAccountRaw } from "./account";
 export * from "./nft";
 export * from "./operation";
+export * from "./transaction";
