+++ libs/ledger-live-common/src/families/stacks/transaction.ts
@@ -3,13 +3,13 @@ import { StacksNetwork } from "./bridge/utils/api.types";
 import type { Account } from "@ledgerhq/types-live";
 
 import type { Transaction, TransactionRaw } from "./types";
+import { formatTransactionStatus } from "@ledgerhq/coin-framework/formatters";
 import {
-  formatTransactionStatusCommon as formatTransactionStatus,
   fromTransactionCommonRaw,
   fromTransactionStatusRawCommon as fromTransactionStatusRaw,
   toTransactionCommonRaw,
   toTransactionStatusRawCommon as toTransactionStatusRaw,
-} from "@ledgerhq/coin-framework/transaction/common";
+} from "@ledgerhq/coin-framework/serialization";
 import { getAccountUnit } from "../../account";
 import { formatCurrencyUnit } from "../../currencies";
